:sparkles: Coup de norminette on lexer
*
