+++ src/builtins/ms_cd/navigate_2.c
@@ -6,7 +6,7 @@
 /*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/13 17:43:51 by owalsh            #+#    #+#             */
-/*   Updated: 2022/10/06 13:47:53 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/10/06 14:21:04 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,9 +89,8 @@ int	nav_backwards(char *path, char **env)
 		free(pwd);
 		path += 3;
 	}
-	if (!path || !path[0])
-		return (EXIT_SUCCESS);
-	else if (!ft_strcmp(path, ".") || !ft_strcmp(path, "./"))
+	if (!path || !path[0]
+		|| !ft_strcmp(path, ".") || !ft_strcmp(path, "./"))
 		return (EXIT_SUCCESS);
 	pwd = ft_getenv("PWD");
 	if (!ft_strncmp(path, "..", 3))
